requirements.txt
apscheduler>=3.6.0
configparser>=3.5.0
dateparser>=0.7.2
flake8>=3.5.0
flask>=0.12.2
flask-api>=1.1b2
flask-cors>=3.0.8 
flask-debugtoolbar>=0.10.1
flask-migrate>=2.1.1
flask-script>=2.0.6
flask-sqlalchemy>=2.3.2
gunicorn>=20.0.4
isort>=4.3.4
mypy>=0.720
pendulum>=2.0.4
pillow>=7.0.0
pytz>=2016.7
requests>=2.18.4
returns>=0.12.0
six>=1.10.0
slackclient>=2.5.0
sqlalchemy>=1.3.5
werkzeug>=0.12